break up schema
