+++ guardrails/schema/json_schema.py
@@ -1,18 +1,6 @@
 import json
-import pprint
 from copy import deepcopy
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Dict,
-    List,
-    Optional,
-    Sequence,
-    Set,
-    Tuple,
-    Type,
-    Union,
-)
+from typing import Any, Dict, List, Optional, Tuple, Type, Union
 
 from lxml import etree as ET
 from pydantic import BaseModel
@@ -20,7 +8,7 @@ from typing_extensions import Self
 
 from guardrails import validator_service
 from guardrails.classes.history import Iteration
-from guardrails.datatypes import Choice, DataType, Object, String
+from guardrails.datatypes import Choice, DataType, Object
 from guardrails.llm_providers import (
     AsyncOpenAICallable,
     AsyncOpenAIChatCallable,
@@ -30,6 +18,7 @@ from guardrails.llm_providers import (
 )
 from guardrails.logger import logger
 from guardrails.prompt import Instructions, Prompt
+from guardrails.schema.schema import Schema
 from guardrails.utils.constants import constants
 from guardrails.utils.json_utils import (
     extract_json_from_ouput,
@@ -46,182 +35,7 @@ from guardrails.utils.reask_utils import (
     prune_obj_for_reasking,
 )
 from guardrails.utils.telemetry_utils import trace_validation_result
-from guardrails.validator_base import (
-    FailResult,
-    ValidatorSpec,
-    check_refrain_in_dict,
-    filter_in_dict,
-)
-
-if TYPE_CHECKING:
-    pass
-
-
-class Schema:
-    """Schema class that holds a _schema attribute."""
-
-    reask_prompt_vars: Set[str]
-
-    def __init__(
-        self,
-        schema: DataType,
-        reask_prompt_template: Optional[str] = None,
-        reask_instructions_template: Optional[str] = None,
-    ) -> None:
-        self.root_datatype = schema
-
-        # Setup reask templates
-        self.reask_prompt_template = reask_prompt_template
-        self.reask_instructions_template = reask_instructions_template
-
-    @classmethod
-    def from_xml(
-        cls,
-        root: ET._Element,
-        reask_prompt_template: Optional[str] = None,
-        reask_instructions_template: Optional[str] = None,
-    ) -> Self:
-        """Create a schema from an XML element."""
-        raise NotImplementedError
-
-    def __repr__(self) -> str:
-        # FIXME make sure this is pretty
-        return f"{self.__class__.__name__}({pprint.pformat(self.root_datatype)})"
-
-    @property
-    def reask_prompt_template(self) -> Optional[Prompt]:
-        return self._reask_prompt_template
-
-    @reask_prompt_template.setter
-    def reask_prompt_template(self, value: Optional[str]) -> None:
-        self.check_valid_reask_prompt(value)
-        if value is not None:
-            self._reask_prompt_template = Prompt(value)
-        else:
-            self._reask_prompt_template = None
-
-    @property
-    def reask_instructions_template(self) -> Optional[Instructions]:
-        return self._reask_instructions_template
-
-    @reask_instructions_template.setter
-    def reask_instructions_template(self, value: Optional[str]) -> None:
-        if value is not None:
-            self._reask_instructions_template = Instructions(value)
-        else:
-            self._reask_instructions_template = None
-
-    def validate(
-        self,
-        iteration: Iteration,
-        data: Any,
-        metadata: Dict,
-        attempt_number: int = 0,
-        **kwargs,
-    ) -> Any:
-        """Validate a dictionary of data against the schema.
-
-        Args:
-            data: The data to validate.
-
-        Returns:
-            The validated data.
-        """
-        raise NotImplementedError
-
-    async def async_validate(
-        self, iteration: Iteration, data: Any, metadata: Dict, attempt_number: int = 0
-    ) -> Any:
-        """Asynchronously validate a dictionary of data against the schema.
-
-        Args:
-            data: The data to validate.
-
-        Returns:
-            The validated data.
-        """
-        raise NotImplementedError
-
-    def transpile(self, method: str = "default") -> str:
-        """Convert the XML schema to a string that is used for prompting a
-        large language model.
-
-        Returns:
-            The prompt.
-        """
-        raise NotImplementedError
-
-    def parse(self, output: str, **kwargs) -> Tuple[Any, Optional[Exception]]:
-        """Parse the output from the large language model.
-
-        Args:
-            output: The output from the large language model.
-
-        Returns:
-            The parsed output, and the exception that was raised (if any).
-        """
-        raise NotImplementedError
-
-    def introspect(
-        self, data: Any
-    ) -> Tuple[Sequence[ReAsk], Optional[Union[str, Dict]]]:
-        """Inspect the data for reasks.
-
-        Args:
-            data: The data to introspect.
-
-        Returns:
-            A list of ReAsk objects.
-        """
-        raise NotImplementedError
-
-    def get_reask_setup(
-        self,
-        reasks: Sequence[ReAsk],
-        original_response: Any,
-        use_full_schema: bool,
-        prompt_params: Optional[Dict[str, Any]] = None,
-    ) -> Tuple["Schema", Prompt, Instructions]:
-        """Construct a schema for reasking, and a prompt for reasking.
-
-        Args:
-            reasks: List of tuples, where each tuple contains the path to the
-                reasked element, and the ReAsk object (which contains the error
-                message describing why the reask is necessary).
-            original_response: The value that was returned from the API, with reasks.
-            use_full_schema: Whether to use the full schema, or only the schema
-                for the reasked elements.
-
-        Returns:
-            The schema for reasking, and the prompt for reasking.
-        """
-        raise NotImplementedError
-
-    def preprocess_prompt(
-        self,
-        prompt_callable: PromptCallableBase,
-        instructions: Optional[Instructions],
-        prompt: Prompt,
-    ):
-        """Preprocess the instructions and prompt before sending it to the
-        model.
-
-        Args:
-            prompt_callable: The callable to be used to prompt the model.
-            instructions: The instructions to preprocess.
-            prompt: The prompt to preprocess.
-        """
-        raise NotImplementedError
-
-    def check_valid_reask_prompt(self, reask_prompt: Optional[str]) -> None:
-        if reask_prompt is None:
-            return
-
-        # Check that the reask prompt has the correct variables
-
-        # TODO decide how to check this
-        # variables = get_template_variables(reask_prompt)
-        # assert set(variables) == self.reask_prompt_vars
+from guardrails.validator_base import FailResult, check_refrain_in_dict, filter_in_dict
 
 
 class JsonSchema(Schema):
@@ -640,256 +454,6 @@ class JsonSchema(Schema):
     def transpile(self, method: str = "default") -> str:
         transpiler = getattr(Schema2Prompt, method)
         return transpiler(self)
-
-
-class StringSchema(Schema):
-    reask_prompt_vars = {"previous_response", "output_schema", "error_messages"}
-
-    def __init__(
-        self,
-        schema: String,
-        reask_prompt_template: Optional[str] = None,
-        reask_instructions_template: Optional[str] = None,
-    ) -> None:
-        super().__init__(
-            schema,
-            reask_prompt_template=reask_prompt_template,
-            reask_instructions_template=reask_instructions_template,
-        )
-        self.root_datatype = schema
-
-    @classmethod
-    def from_xml(
-        cls,
-        root: ET._Element,
-        reask_prompt_template: Optional[str] = None,
-        reask_instructions_template: Optional[str] = None,
-    ) -> Self:
-        if len(root) != 0:
-            raise ValueError("String output schemas must not have children.")
-
-        strict = False
-        if "strict" in root.attrib and root.attrib["strict"] == "true":
-            strict = True
-
-        schema = String.from_xml(root, strict=strict)
-
-        return cls(
-            schema=schema,
-            reask_prompt_template=reask_prompt_template,
-            reask_instructions_template=reask_instructions_template,
-        )
-
-    @classmethod
-    def from_string(
-        cls,
-        validators: Sequence[ValidatorSpec],
-        description: Optional[str] = None,
-        reask_prompt_template: Optional[str] = None,
-        reask_instructions_template: Optional[str] = None,
-    ):
-        strict = False
-
-        schema = String.from_string_rail(
-            validators, description=description, strict=strict
-        )
-
-        return cls(
-            schema=schema,
-            reask_prompt_template=reask_prompt_template,
-            reask_instructions_template=reask_instructions_template,
-        )
-
-    def get_reask_setup(
-        self,
-        reasks: List[FieldReAsk],
-        original_response: FieldReAsk,
-        use_full_schema: bool,
-        prompt_params: Optional[Dict[str, Any]] = None,
-    ) -> Tuple[Schema, Prompt, Instructions]:
-        pruned_tree_string = self.transpile()
-
-        reask_prompt_template = self.reask_prompt_template
-        if reask_prompt_template is None:
-            reask_prompt_template = Prompt(
-                constants["high_level_string_reask_prompt"]
-                + constants["complete_string_suffix"]
-            )
-
-        error_messages = "\n".join(
-            [
-                f"- {fail_result.error_message}"
-                for reask in reasks
-                for fail_result in reask.fail_results
-            ]
-        )
-
-        prompt = reask_prompt_template.format(
-            previous_response=original_response.incorrect_value,
-            error_messages=error_messages,
-            output_schema=pruned_tree_string,
-            **(prompt_params or {}),
-        )
-
-        instructions = self.reask_instructions_template
-        if instructions is None:
-            instructions = Instructions("You are a helpful assistant.")
-        instructions = instructions.format(**(prompt_params or {}))
-
-        return self, prompt, instructions
-
-    def parse(self, output: str, **kwargs) -> Tuple[Any, Optional[Exception]]:
-        return output, None
-
-    def validate(
-        self,
-        iteration: Iteration,
-        data: Any,
-        metadata: Dict,
-        **kwargs,
-    ) -> Any:
-        """Validate a dictionary of data against the schema.
-
-        Args:
-            data: The data to validate.
-
-        Returns:
-            The validated data.
-        """
-        if data is None:
-            return None
-
-        if not isinstance(data, str):
-            raise TypeError(f"Argument `data` must be a string, not {type(data)}.")
-
-        # FIXME instead of writing the validation infrastructure for dicts (JSON),
-        #  make it more structure-invariant
-        dummy_key = "string"
-        validation = self.root_datatype.collect_validation(
-            key=dummy_key,
-            value=data,
-            schema={
-                dummy_key: data,
-            },
-        )
-
-        validated_response, metadata = validator_service.validate(
-            value=data,
-            metadata=metadata,
-            validator_setup=validation,
-            iteration=iteration,
-        )
-
-        validated_response = {dummy_key: validated_response}
-
-        if check_refrain_in_dict(validated_response):
-            # If the data contains a `Refain` value, we return an empty
-            # dictionary.
-            logger.debug("Refrain detected.")
-            validated_response = {}
-
-        # Remove all keys that have `Filter` values.
-        validated_response = filter_in_dict(validated_response)
-
-        if dummy_key in validated_response:
-            return validated_response[dummy_key]
-        return None
-
-    async def async_validate(
-        self,
-        iteration: Iteration,
-        data: Any,
-        metadata: Dict,
-    ) -> Any:
-        """Validate a dictionary of data against the schema.
-
-        Args:
-            data: The data to validate.
-
-        Returns:
-            The validated data.
-        """
-        if data is None:
-            return None
-
-        if not isinstance(data, str):
-            raise TypeError(f"Argument `data` must be a string, not {type(data)}.")
-
-        dummy_key = "string"
-        validation = self.root_datatype.collect_validation(
-            key=dummy_key,
-            value=data,
-            schema={
-                dummy_key: data,
-            },
-        )
-
-        validated_response, metadata = await validator_service.async_validate(
-            value=data,
-            metadata=metadata,
-            validator_setup=validation,
-            iteration=iteration,
-        )
-
-        validated_response = {dummy_key: validated_response}
-
-        if check_refrain_in_dict(validated_response):
-            # If the data contains a `Refain` value, we return an empty
-            # dictionary.
-            logger.debug("Refrain detected.")
-            validated_response = {}
-
-        # Remove all keys that have `Filter` values.
-        validated_response = filter_in_dict(validated_response)
-
-        if dummy_key in validated_response:
-            return validated_response[dummy_key]
-        return None
-
-    def introspect(
-        self, data: Union[ReAsk, Optional[str]]
-    ) -> Tuple[List[FieldReAsk], Optional[str]]:
-        if isinstance(data, FieldReAsk):
-            return [data], None
-        return [], data  # type: ignore
-
-    def preprocess_prompt(
-        self,
-        prompt_callable: PromptCallableBase,
-        instructions: Optional[Instructions],
-        prompt: Prompt,
-    ):
-        if isinstance(prompt_callable, OpenAICallable) or isinstance(
-            prompt_callable, AsyncOpenAICallable
-        ):
-            prompt.source += "\n\nString Output:\n\n"
-        if (
-            isinstance(prompt_callable, OpenAIChatCallable)
-            or isinstance(prompt_callable, AsyncOpenAIChatCallable)
-        ) and not instructions:
-            instructions = Instructions(
-                "You are a helpful assistant, expressing yourself through a string."
-            )
-
-        return instructions, prompt
-
-    def transpile(self, method: str = "default") -> str:
-        obj = self.root_datatype
-        schema = ""
-        if obj.description is not None:
-            schema += (
-                "Here's a description of what I want you to generate: "
-                f"{obj.description}"
-            )
-        if not obj.validators_attr.empty:
-            schema += (
-                "\n\nYour generated response should satisfy the following properties:"
-            )
-            for validator in obj.validators_attr.validators:
-                schema += f"\n- {validator.to_prompt()}"
-
-        schema += "\n\nDon't talk; just go."
-        return schema
 
 
 class Schema2Prompt:
